note: bugs fixed.
Fixed errors when reading the lock from a file and when writing to a file. Removed the "Save" button when editing notes, which caused incorrect work. Instead, just enter the desired note, click on "Enter" and close the editing window.
We also fixed an error in which, while editing a note, its name was still reset.